requirements.txt
discord.py
ffmpeg
python-dotenv
